完成余下屎山

+++ forum-api/src/main/java/top/vs/forum/po/Star.java
@@ -1,11 +1,12 @@
 package top.vs.forum.po;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.extension.activerecord.Model;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.extension.activerecord.Model;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -35,7 +36,7 @@ public class Star extends Model<Star> {
     /**
      * 被收藏帖子id
      */
-    private Integer starId;
+    private Long starId;
 
 
     @Override
